% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{CreateWaiterScreen}
\alias{CreateWaiterScreen}
\title{CreateWaiterScreen}
\usage{
CreateWaiterScreen(ID)
}
\arguments{
\item{ID}{\code{string}}
}
\value{
Waiter object
}
\description{
CreateWaiterScreen
}
