% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameToHtmlTable.R
\name{DataFrameToHtmlTable}
\alias{DataFrameToHtmlTable}
\title{DataFrameToHtmlTable}
\usage{
DataFrameToHtmlTable(
  DataFrame,
  SemanticTableClass = "ui celled table",
  RotatedHeaderNames = character()
)
}
\arguments{
\item{RotatedHeaderNames}{}
}
\description{
DataFrameToHtmlTable
}
\author{
Bastian Reiter
}
