% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineGridTemplate.R
\name{DefineGridTemplate}
\alias{DefineGridTemplate}
\title{DefineGridTemplate}
\usage{
DefineGridTemplate()
}
\value{
The output of shiny.semantic::grid_template()
}
\description{
Define grid layout for desktop and mobile clients using shiny.semantic::grid_template()
}
\keyword{internal}
