% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainUIGridTemplate.R
\name{MainUIGridTemplate}
\alias{MainUIGridTemplate}
\title{MainUIGridTemplate}
\usage{
MainUIGridTemplate()
}
\value{
The output of shiny.semantic::grid_template()
}
\description{
Define main UI grid layout for desktop and mobile clients using shiny.semantic::grid_template()
}
\keyword{internal}
