% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAutonomousApp.R
\name{RunAutonomousApp}
\alias{RunAutonomousApp}
\title{RunAutonomousApp}
\usage{
RunAutonomousApp(
  ShinyApp,
  AppArguments = NULL,
  Host = "127.0.0.1",
  Port = 49152
)
}
\arguments{
\item{ShinyApp}{}
}
\value{
An \code{r_process} object. See also documentation for \code{callr::r_bg()}.
}
\description{
Set up a background process essentially running a separate R session. Use this session to run a Shiny app autonomously from hosting R session.
}
\author{
Bastian Reiter
}
