% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServerExplorer.R
\name{ServerExplorer}
\alias{ServerExplorer}
\title{ServerExplorer}
\usage{
ServerExplorer(ServerWorkspaceInfo = NULL, DSConnections = NULL)
}
\arguments{
\item{DSConnections}{\code{list} of \code{DSConnection} objects}
}
\description{
Launch Module \code{ModServerExplorer} as a Shiny app.
}
\author{
Bastian Reiter
}
