% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameToHtmlTable.R
\name{DataFrameToHtmlTable}
\alias{DataFrameToHtmlTable}
\title{DataFrameToHtmlTable}
\usage{
DataFrameToHtmlTable(
  DataFrame,
  TableID = NULL,
  CategoryColumn = NULL,
  CellClassColumns = NULL,
  ColContentHorizontalAlign = "left",
  ColumnClass = NULL,
  ColumnLabels = NULL,
  RotatedHeaderNames = character(),
  RowColorColumn = NULL,
  SemanticTableClass = "ui celled table",
  TurnLogicalIntoIcon = FALSE
)
}
\arguments{
\item{DataFrame}{\code{data.frame} or \code{tibble}}

\item{TableID}{\code{string} Used to tidentify the table object in the DOM}

\item{CategoryColumn}{\code{string}}

\item{CellClassColumns}{character vector}

\item{ColContentHorizontalAlign}{Either a single string for all columns or a named character vector to determine horizontal content alignment for specific columns}

\item{ColumnClass}{Either a single string for all columns or a named character vector to determine table cell classes for specific columns}

\item{ColumnLabels}{named character vector}

\item{RowColorColumn}{\code{string}}

\item{TurnLogicalIntoIcon}{}
}
\value{
HTML code
}
\description{
DataFrameToHtmlTable
}
\author{
Bastian Reiter
}
