% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameToHtmlTable.R
\name{DataFrameToHtmlTable}
\alias{DataFrameToHtmlTable}
\title{DataFrameToHtmlTable}
\usage{
DataFrameToHtmlTable(
  DataFrame,
  TableID = NULL,
  CategoryColumn = NULL,
  CellClassColumns = NULL,
  ColContentHorizontalAlign = "left",
  ColumnClass = NULL,
  ColumnIcons = NULL,
  ColumnLabels = NULL,
  ColumnLabelsLineBreak = FALSE,
  ColumnMaxWidth = 1000,
  RotatedHeaderNames = character(),
  RowColorColumn = NULL,
  SemanticTableClass = "ui celled table",
  TableStyle = "",
  TurnLogicalIntoIcon = FALSE,
  TurnColorValuesIntoDots = FALSE
)
}
\arguments{
\item{DataFrame}{\code{data.frame} or \code{tibble}}

\item{TableID}{\code{string} Used to tidentify the table object in the DOM}

\item{CategoryColumn}{\code{string}}

\item{CellClassColumns}{character vector}

\item{ColContentHorizontalAlign}{Either a single string for all columns or a named character vector to determine horizontal content alignment for specific columns}

\item{ColumnClass}{Either a single string for all columns or a named character vector to determine table cell classes for specific columns}

\item{ColumnIcons}{named \code{vector}}

\item{ColumnLabels}{named \code{vector}}

\item{ColumnLabelsLineBreak}{\code{logical} Indicating whether column labels should span across two lines}

\item{ColumnMaxWidth}{\code{integer} Maximum width of columns in character spaces}

\item{RotatedHeaderNames}{\code{vector}}

\item{RowColorColumn}{\code{string}}

\item{SemanticTableClass}{\code{string}}

\item{TableStyle}{A \code{string} containing CSS style elements}

\item{TurnLogicalIntoIcon}{\code{logical}}

\item{TurnColorValuesIntoDots}{\code{logical}}
}
\value{
HTML code
}
\description{
DataFrameToHtmlTable
}
\author{
Bastian Reiter
}
