% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameToHtmlTable.R
\name{DataFrameToHtmlTable}
\alias{DataFrameToHtmlTable}
\title{DataFrameToHtmlTable}
\usage{
DataFrameToHtmlTable(
  DataFrame,
  CategoryColumn = NULL,
  CellClassColumns = NULL,
  ColumnClass = NULL,
  ColumnLabels = NULL,
  RotatedHeaderNames = character(),
  RowColorColumn = NULL,
  SemanticTableClass = "ui celled table",
  TurnLogicalIntoIcon = FALSE
)
}
\arguments{
\item{CategoryColumn}{\code{string}}

\item{CellClassColumns}{character vector}

\item{ColumnClass}{Either a single string for all columns or a named character vector to determine table cell classes for specific columns}

\item{ColumnLabels}{named character vector}

\item{RowColorColumn}{\code{string}}

\item{TurnLogicalIntoIcon}{}

\item{ColumnHorizontalAlign}{Either a single string for all columns or a named character vector to determine horizontal content alignment for specific columns}
}
\description{
DataFrameToHtmlTable
}
\author{
Bastian Reiter
}
