% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameToHtmlTable.R
\name{DataFrameToHtmlTable}
\alias{DataFrameToHtmlTable}
\title{DataFrameToHtmlTable}
\usage{
DataFrameToHtmlTable(
  DataFrame,
  SemanticTableClass = "ui celled table",
  RotatedHeaderNames = character(),
  TurnLogicalIntoIcons = FALSE,
  CategoryColumn = NULL,
  RowColorColumn = NULL
)
}
\arguments{
\item{CategoryColumn}{\code{string}}

\item{RowColorColumn}{\code{string}}
}
\description{
DataFrameToHtmlTable
}
\author{
Bastian Reiter
}
