% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StartCCPhosApp.R
\name{StartCCPhosApp}
\alias{StartCCPhosApp}
\title{StartCCPhosApp}
\usage{
StartCCPhosApp(TestData = NULL)
}
\arguments{
\item{CCPConnections}{List of DSConnection objects}

\item{CCPTestData}{List | Optional CCP test data}
}
\description{
Launch Shiny app
}
\author{
Bastian Reiter
}
